Moved optional prompt fields to own tabs in prompt modal
