+++ apps/ui/admin/src/App.scss
@@ -28,3 +28,8 @@ li {
   height: 460px;
   overflow-y: auto;
 }
+
+#prompts-tab-panel-wrapper {
+  height: 460px;
+  overflow-y: auto;
+}
